Improving model

+++ include/network.h
@@ -20,7 +20,7 @@ typedef struct Network
 
 void network_init(Network* net,int neurons_input,int neurons_hidden,int neurons_output);
 void network_free(Network *net);
-void network_predict(Network *net,uint8_t *inputs);
+void network_predict(Network *net,double *inputs);
 
 
 typedef struct Trainer
@@ -32,6 +32,6 @@ typedef struct Trainer
 Trainer* trainer_init(Trainer *trainer,Network *net);
 void trainer_free(Trainer *trainer);
 void trainer_SGD_train(Trainer* trainer, Network* network, double* input, double* output, double lr);
-void trainer_Mini_Batch_train(Trainer *trainer, Network *network, uint8_t **input, uint8_t *output, uint8_t epoch, uint32_t batch_size, double learning_rate, uint32_t  dataset_size);
+void trainer_Mini_Batch_train(Trainer *trainer, Network *network, double **input, uint8_t *output, uint8_t epoch, uint32_t batch_size, double learning_rate, uint32_t  dataset_size);
 //too many parameters , will do for now
 #endif
